requirements.txt
fastapi
uvicorn
pyodbc
dotenv
openai
celery
redis
sentence-transformers
faiss-cpu
numpy
inflect